requirements (1).txt
typer==0.12.5
httpx==0.27.2
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pydantic==2.9.2
rich==13.9.2
